Update 2.4.x to AHC 1.9.33

+++ framework/src/play-java-ws/src/main/java/play/libs/ws/ning/NingWSCookie.java
@@ -45,7 +45,7 @@ public class NingWSCookie implements WSCookie {
     }
 
     public Integer getMaxAge() {
-        return ahcCookie.getMaxAge();
+        return ((int) ahcCookie.getMaxAge());
     }
 
     public Boolean isSecure() {
